requirements.txt
fastapi
uvicorn
python-multipart
twilio
sqlalchemy
psycopg2-binary
google-api-python-client
google-auth
google-auth-oauthlib
httpx
stripe
reportlab
